Rename MutltiExecutionType, improved code quality, and rewrote javadocs
