+++ src/main/java/seedu/address/logic/parser/RemovalBasedCommandExecutionTypeParser.java
@@ -11,102 +11,112 @@ import static seedu.address.logic.util.PrefixUtil.areAnyPrefixesPresent;
 import java.util.List;
 import java.util.function.Predicate;
 
-import seedu.address.commons.core.commandexecutiontype.CommandExecutionType;
+import seedu.address.commons.core.commandexecutiontype.RemovalBasedCommandExecutionType;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.MultiExecutionTypeCommand;
+import seedu.address.logic.commands.RemovalBasedCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.internship.InternshipApplication;
 
 /**
- * Parses input arguments and creates a new DeleteCommand object.
+ * Parses input arguments and creates a new RemovalBasedCommand based on the command execution type.
+ * Mainly checks whether the input is of a valid format and whether it conforms to any of the
+ * command execution type formats.
  */
-public class MultiExecutionTypeCommandParser {
+public class RemovalBasedCommandExecutionTypeParser implements Parser<RemovalBasedCommand> {
 
-    /** Prefixes that are accepted for execution as fields in DeleteCommand. */
+    /** Prefixes that are accepted for execution by field in RemovalBasedCommand. */
     private static final Prefix[] acceptedPrefixes = { PREFIX_COMPANY, PREFIX_ROLE, PREFIX_STATUS };
 
+    /** Holds the specific command string that will be used to create the command. */
+    private final String commandWord;
+
+    public RemovalBasedCommandExecutionTypeParser(String commandWord) {
+        this.commandWord = commandWord;
+    }
+
     /**
-     * Tokenizes the given {@code String} of arguments in the context of the DeleteCommand
+     * Tokenizes the given {@code String} of arguments in the context of the RemovalBasedCommand
      * and retrieves the appropriate execution type based on the content of the string.
      *
-     * @return the appropriate DeleteCommand based the the execution type.
+     * @return the appropriate RemovalBasedCommand based the the execution type.
      * @throws ParseException if the user input does not conform the expected format.
      */
-    public MultiExecutionTypeCommand parse(String args, String commandWord) throws ParseException {
+    public RemovalBasedCommand parse(String args) throws ParseException {
 
         ArgumentMultimap argMultimap = ArgumentTokenizer.tokenize(args, ALL_PREFIXES);
-        CommandExecutionType executionType = CommandExecutionType.getExecutionType(args, argMultimap);
+        RemovalBasedCommandExecutionType executionType =
+            RemovalBasedCommandExecutionType.getExecutionType(args, argMultimap);
 
         switch (executionType) {
         case BY_INDEX:
-            return commandByIndex(args, commandWord);
+            return commandByIndex(args);
         case BY_INDICES:
-            return commandByIndices(args, commandWord);
+            return commandByIndices(args);
         case BY_FIELD:
-            return commandByField(args, commandWord);
+            return commandByField(args);
         default:
             // this should never happen
             assert false;
-            return commandByIndex(args, commandWord);
+            return commandByIndex(args);
         }
 
     }
 
     /**
-     * Parses the given {@code String} of arguments in the context of a DeleteCommand
-     * that is to be executed by index and returns a DeleteCommand object for execution.
+     * Parses the given {@code String} of arguments in the context of a RemovalBasedCommand
+     * that is to be executed by index and returns a RemovalBasedCommand for execution.
      *
      * @param args the argument to be parsed into an Index object.
      * @throws ParseException if the user input does not conform the expected format.
      */
-    public MultiExecutionTypeCommand commandByIndex(String args, String commandWord) throws ParseException {
+    public RemovalBasedCommand commandByIndex(String args) throws ParseException {
         try {
             Index index = ParserUtil.parseIndex(args);
-            return new MultiExecutionTypeCommand(index, CommandExecutionType.BY_INDEX, commandWord);
+            return new RemovalBasedCommand(index, RemovalBasedCommandExecutionType.BY_INDEX, commandWord);
         } catch (ParseException pe) {
             String exceptionMessage = String.format(MESSAGE_INVALID_COMMAND_FORMAT,
-                MultiExecutionTypeCommand.MESSAGE_USAGE_BY_INDEX.apply(commandWord));
+                RemovalBasedCommand.MESSAGE_USAGE_BY_INDEX.apply(commandWord));
             throw new ParseException(exceptionMessage, pe);
         }
     }
 
     /**
-     * Parses the given {@code String} of arguments in the context of the DeleteCommand
-     * that is to be executed by multiple indexes and returns a DeleteCommand object for execution.
+     * Parses the given {@code String} of arguments in the context of the RemovalBasedCommand
+     * that is to be executed by multiple indexes and returns a RemovalBasedCommand for execution.
      *
      * @param args the argument to be parsed into an a set of Index object.
      * @throws ParseException if the user input does not conform to the expected format.
      */
-    public MultiExecutionTypeCommand commandByIndices(String args, String commandWord) throws ParseException {
+    public RemovalBasedCommand commandByIndices(String args) throws ParseException {
         try {
-            List<Index> indicesList = ParserUtil.parseIndices(args, CommandExecutionType.INDICES_DELIMITER);
-            return new MultiExecutionTypeCommand(indicesList, CommandExecutionType.BY_INDICES, commandWord);
+            List<Index> indicesList = ParserUtil.parseIndices(args, RemovalBasedCommandExecutionType.INDICES_DELIMITER);
+            return new RemovalBasedCommand(indicesList, RemovalBasedCommandExecutionType.BY_INDICES, commandWord);
         } catch (ParseException pe) {
             String exceptionMessage = String.format(MESSAGE_INVALID_COMMAND_FORMAT,
-                MultiExecutionTypeCommand.MESSAGE_USAGE_BY_INDICES.apply(commandWord));
+                RemovalBasedCommand.MESSAGE_USAGE_BY_INDICES.apply(commandWord));
             throw new ParseException(exceptionMessage, pe);
         }
     }
 
     /**
-     * Parses the given {@code String} of arguments in the context of the DeleteCommand
-     * that is to be executed by accepted fields/prefixes and returns a DeleteCommand object for execution.
+     * Parses the given {@code String} of arguments in the context of the RemovalBasedCommand
+     * that is to be executed by accepted fields/prefixes and returns a RemovalBasedCommand for execution.
      *
      * @throws ParseException if the user input does not conform to the expected format.
      */
-    public MultiExecutionTypeCommand commandByField(String args, String commandWord) throws ParseException {
+    public RemovalBasedCommand commandByField(String args) throws ParseException {
         ArgumentMultimap argMultimap = ArgumentTokenizer.tokenize(args, acceptedPrefixes);
         boolean hasOnlyOneField = (argMultimap.getSize() - 1) == 1;
         boolean hasAcceptedPrefixesPresent = areAnyPrefixesPresent(argMultimap, acceptedPrefixes);
         boolean isValidField = hasOnlyOneField & hasAcceptedPrefixesPresent;
         if (!isValidField) {
             String exceptionMessage = String.format(MESSAGE_INVALID_COMMAND_FORMAT,
-                MultiExecutionTypeCommand.MESSAGE_USAGE_BY_FIELD.apply(commandWord));
+                RemovalBasedCommand.MESSAGE_USAGE_BY_FIELD.apply(commandWord));
             throw new ParseException(exceptionMessage);
         }
         Predicate<InternshipApplication> predicate = getFieldPredicate(argMultimap, acceptedPrefixes);
 
-        return new MultiExecutionTypeCommand(predicate, CommandExecutionType.BY_FIELD, commandWord);
+        return new RemovalBasedCommand(predicate, RemovalBasedCommandExecutionType.BY_FIELD, commandWord);
     }
 
 }
